requirements.txt
# ── Core ──────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0.0

# ── Neuroscience / Imaging ─────────────────────────────
mne>=1.5.0                   # EEG loading + PhysioNet download
nilearn>=0.10.0              # fMRI Haxby dataset
nibabel>=5.0.0               # NIfTI I/O

# ── Data / Vision ──────────────────────────────────────
datasets>=2.14.0             # HuggingFace FER2013
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# ── Experiment tracking ────────────────────────────────
mlflow>=2.5.0

# ── API ────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ── Dev / Testing ──────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0                # for FastAPI test client
